requirements.txt
scapy>=2.6.0
fastapi>=0.115.0
uvicorn>=0.32.0
pyyaml>=6.0.2
pydantic>=2.10.0
python-multipart>=0.0.9
cryptography>=44.0.0
asyncssh>=2.14.0
netmiko>=4.4.0
